requirements.txt
discord.py
python-dotenv==0.13.0
